+++ app/auth/actions.ts
@@ -27,7 +27,7 @@ export async function signInWithGoogle() {
   }
 }
 
-export async function signOut() {
+export async function logout() {
   const supabase = await createClient();
 
   const { error } = await supabase.auth.signOut();
